chore(fastify): Refactor tests and add exports.test
