+++ packages/fastify/src/exports.test.ts
@@ -1,0 +1,7 @@
+import * as publicExports from './index';
+
+describe('/api public exports', () => {
+  it('should not include a breaking change', () => {
+    expect(Object.keys(publicExports).sort()).toMatchSnapshot();
+  });
+});
